refactor: move Authentication error type definitions into domain
