+++ src/services/kratos/private.ts
@@ -1,7 +1,9 @@
 import { getKratosConfig } from "@config"
+import {
+  MissingExpiredAtKratosError,
+  UnknownKratosError,
+} from "@domain/authentication/errors"
 import { Configuration, V0alpha2Api, V0alpha2ApiInterface } from "@ory/client"
-
-import { MissingExpiredAtKratosError, UnknownKratosError } from "./errors"
 
 const { publicApi, adminApi } = getKratosConfig()
